feat (intelligent alerts): added api to update evaluation metrics from Kusto cluster

+++ src/service/Domain/Domain/ValueObjects/PerformanceMetric.cs
@@ -2,11 +2,11 @@
 {
     public class PerformanceMetric
     {
-        public int Percentile95Latency { get; private set; }
-        public int Percentile90Latency { get; private set; }
-        public int AverageLatency { get; private set; }
+        public double Percentile95Latency { get; private set; }
+        public double Percentile90Latency { get; private set; }
+        public double AverageLatency { get; private set; }
 
-        public PerformanceMetric(int p95, int p90, int averge)
+        public PerformanceMetric(double p95, double p90, double averge)
         {
             Percentile95Latency = p95;
             Percentile90Latency = p90;
